Expose oak containers launcher as library
